backend/requirements.txt
Django>=4.2.0,<4.3.0
mysqlclient
djangorestframework>=3.14.0,<3.15.0
djangorestframework-simplejwt
PyJWT
pymysql
django-cors-headers>=4.0.0,<4.1.0
Pillow>=10.0.0,<10.1.0
psycopg2-binary>=2.9.5,<2.10.0
python-dotenv>=1.0.0,<1.1.0
gunicorn>=21.0.0 